Add tractorgen-animator, by Peter Miller
This is a major enhancement to the suite of tools available for the
study of ASCII tractor mechanics. It was first submitted by Peter Miller
on the 11th of December 2008, and has undergone rigorous testing on a
daily basis in the interim.

In its simplest form, Mr. Miller suggests the following usage:

    tractorgen | tractorgen-animator

Which will generate a standard ASCII tractor of seven wheels, and
animate its motion across the display terminal.

This serves the very useful purpose of enabling the use of an ASCII
tractor to appropriately remove extant error messages from a user's
terminal. Mr. Miller has employed this method in the reboot procedure of
an embedded device, wherein the ASCII tractor performs the important
task of cleansing the display of unsightly text.

I am personally very fond of the following apparatus:

    for i in `seq 1 100000`; do tractorgen $i | tractorgen-animator; done

Which will attempt to create a succession of all ASCII tractors of wheel
lengths from 1 to 100000 and animate the resulting successful instances
thereof. Apart from being an enthralling method of spending one's
weekend, this provides a "hands-off" experimental routine which is
surely welcome in these days of high-pressure research.